codes.ordered <- .Alias(as.integer), not as.numeric
git-svn-id: https://svn.r-project.org/R/branches/R-0-62-patches@1949 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/base/R/factor.R
@@ -41,7 +41,7 @@ codes.factor <- function(x)
   rank(levels(x))[x]
 }
 
-codes.ordered <- .Alias(as.numeric)
+codes.ordered <- .Alias(as.integer)
 
 "codes<-" <- function(x, value)
 {
